requirements.txt
streamlit
gTTS
easyocr
opencv-python-headless
numpy
Pillow
streamlit-local-storage